# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketing APIs",
        "description": "Marketing kampaniyaları və segment idarəetmə API-si",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/unsubscribe/{userId}/{campaignId}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Abunəlikdən çıxma",
                "description": "İmza ilə təsdiqlənmiş unsubscribe linki.",
                "operationId": "unsubscribe",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 614
                        }
                    },
                    {
                        "name": "campaignId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 8
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "442e4376ce7c22e33dd844d60d62beacda7dc7b985851e9062e4d8efc90af56d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Abunəlikdən uğurla çıxarıldınız",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Yanlış imza"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}